+++ src/components/Inputs/ChatInput/index.ts
@@ -1,0 +1,9 @@
+import LbChatInput from './LbChatInput.vue'
+
+export default LbChatInput
+export type { 
+  ChatInputMenuItem,
+  LbChatInputProps,
+  LbChatInputSlots,
+  LbChatInputEmits
+} from './LbChatInput.vue.d'